Add files via upload
This repository contains the code and analysis performed for AtliQ Hotels, a luxury hotel chain with operations in major Indian cities including Mumbai, Delhi, Hyderabad, and Bangalore. The project focuses on analyzing business data from May to August 2022 to uncover trends and insights that can inform strategic decisions in the hospitality industry.